+++ CHANGELOG.md
@@ -7,15 +7,19 @@
 * Default resync mode for `FlexibleSyncConfiguration` is changed from `manual` to `recoverOrDiscard`. In this mode Realm attempts to recover unsynced local changes and if that fails, then the changes are discarded.(PR [#925](https://github.com/realm/realm-dart/pull/925))
 
 ### Enhancements
+* Added support for asynchronous transactions. (Issue [#802](https://github.com/realm/realm-dart/issues/802))
+  * Added `Transaction` which is a class that exposes an API for committing and rolling back an active transaction.
+  * Added `realm.beginWriteAsync` which returns a `Future<Transaction>` that resolves when the write lock has been obtained.
+  * Added `realm.writeAsync` which opens an asynchronous transaction, invokes the provided callback, then commits the transaction asynchronously.
 
 ### Fixed
+* Added more validations when using `User.apiKeys` to return more meaningful errors when the user cannot perform API key actions - e.g. when the user has been logged in with API key credentials or when the user has been logged out. (Issue [#950](https://github.com/realm/realm-dart/issues/950))
 
 ### Compatibility
 * Realm Studio: 12.0.0 or later.
 
 ### Internal
 * Uses Realm Core v12.9.0
-
 
 ## 0.5.0+beta (2022-10-10)
 
@@ -45,7 +49,7 @@
     }
 
     if (oldSchemaVersion == 3) {
-      final oldPeople = migration.oldRealm.dynamic.all('Person');
+      final oldPeople = migration.oldRealm.all('Person');
       for (final oldPerson in oldPeople) {
         final newPerson = migration.findInNewRealm<Person>(oldPerson);
         if (newPerson == null) {