+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ### Fixed
 * Added more validations when using `User.apiKeys` to return more meaningful errors when the user cannot perform API key actions - e.g. when the user has been logged in with API key credentials or when the user has been logged out. (Issue [#950](https://github.com/realm/realm-dart/issues/950))
-* Fixed `dart run realm_dart generate` and `flutter pub run realm generate` commands to exit with an error code when generation fails. Execution is now terminated when the command is run in a batch with other commands.
+* Fixed `dart run realm_dart generate` and `flutter pub run realm generate` commands to exit with the correct error code on failure.
 
 ### Compatibility
 * Realm Studio: 12.0.0 or later.